+++ router-core/src/app/mod.rs
@@ -20,6 +20,6 @@
 //! The implementations in this module build on the lower-level system components
 //! to provide the actual gateway and proxy behavior defined by user configuration.
 
-pub mod proxy;
 pub mod proxy_fast;
-pub mod gateway;
+pub mod proxy_sni;
+pub mod proxy_host;
