+++ router-core/src/system/writer/mod.rs
@@ -23,24 +23,24 @@ use mapper::{setup_standard_logging, setup_tag_based_logging};
 ///    and initializes the standard `env_logger` directly as a last resort,
 ///    logging a warning about using the default configuration.
 pub fn writer_start() {
-    // Try the tag-based logging first
-    if setup_tag_based_logging().is_ok() {
-        log::info!("Tag-based logging initialized successfully");
-        return;
-    }
+    // // Try the tag-based logging first
+    // if setup_tag_based_logging().is_ok() {
+    //     log::info!("Tag-based logging initialized successfully");
+    //     return;
+    // }
     
-    // Log failure to standard error as the logging system isn't fully up yet.
-    eprintln!("Failed to initialize tag-based logging");
+    // // Log failure to standard error as the logging system isn't fully up yet.
+    // eprintln!("Failed to initialize tag-based logging");
     
-    // Fall back to standard logging (currently env_logger)
-    if setup_standard_logging().is_ok() {
-        // This log might go to stderr depending on env_logger config.
-        log::info!("Standard env_logger-based logging initialized successfully");
-        return;
-    }
+    // // Fall back to standard logging (currently env_logger)
+    // if setup_standard_logging().is_ok() {
+    //     // This log might go to stderr depending on env_logger config.
+    //     log::info!("Standard env_logger-based logging initialized successfully");
+    //     return;
+    // }
     
-    // Log failure to standard error.
-    eprintln!("Failed to initialize standard logging");
+    // // Log failure to standard error.
+    // eprintln!("Failed to initialize standard logging");
     
     // Last resort: standard env_logger to stderr
     // Ensure RUST_LOG is set for env_logger.
